@@ -5,6 +5,6 @@
         /// <summary>
         /// Gets value to put in the DB Create/Update Audit Field
         /// </summary>
-        string GetDbUsername();
+        string? GetDbUsername();
     }
 }
