@@ -10,7 +10,7 @@ namespace OLT.Core
         /// <remarks>
         /// Claim <see cref="OltClaimTypes.GivenName"/>
         /// </remarks>
-        string First { get; set; }
+        string? First { get; set; }
 
         /// <summary>
         /// First Name
@@ -18,7 +18,7 @@ namespace OLT.Core
         /// <remarks>
         /// Claim <see cref="OltClaimTypes.MiddleName"/>
         /// </remarks>
-        string Middle { get; set; }
+        string? Middle { get; set; }
 
         /// <summary>
         /// First Name
@@ -26,7 +26,7 @@ namespace OLT.Core
         /// <remarks>
         /// Claim <see cref="OltClaimTypes.FamilyName"/>
         /// </remarks>
-        string Last { get; set; }
+        string? Last { get; set; }
 
         /// <summary>
         /// Name Suffix (Jr, Sr, II, III, IV, V)
@@ -34,7 +34,7 @@ namespace OLT.Core
         /// <remarks>
         /// Included with <see cref="Last"/> in claim <see cref="OltClaimTypes.FamilyName"/> 
         /// </remarks>
-        string Suffix { get; set; }
+        string? Suffix { get; set; }
 
         /// <summary>
         /// Full Name using <see cref="First"/> <see cref="Middle"/> <see cref="Last"/> <see cref="Suffix"/>
@@ -42,6 +42,6 @@ namespace OLT.Core
         /// <remarks>
         /// Claim <see cref="OltClaimTypes.Name"/>
         /// </remarks>
-        string FullName { get; }
+        string? FullName { get; }
     }
 }
