@@ -6,6 +6,6 @@
     /// </summary>
     public interface IOltValidationError
     {
-        string Message { get; set; }
+        string? Message { get; set; }
     }
 }
