@@ -1,15 +1,84 @@
 ﻿using FluentAssertions;
+using FluentAssertions.Common;
+using Microsoft.AspNetCore;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 using OLT.Constants;
 using OLT.Core;
 using OLT.Logging.Serilog;
 using Serilog;
 using Serilog.Events;
+using Serilog.Extensions.Hosting;
+using Serilog.Extensions.Logging;
+using Seriolog;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using Xunit;
+
+namespace Seriolog
+{
+    /// <summary>
+    /// I took this from the original deprecated source in 7.0
+    /// </summary>
+    public static class SerilogWebHostBuilderExtensions
+    {
+        public static IWebHostBuilder UseSerilog_local(this IWebHostBuilder builder, Serilog.ILogger? logger = null, bool dispose = false, LoggerProviderCollection? providers = null)
+        {
+            LoggerProviderCollection providers2 = providers;
+            Serilog.ILogger logger2 = logger;
+            if (builder == null)
+            {
+                throw new ArgumentNullException("builder");
+            }
+
+            builder.ConfigureServices(delegate (IServiceCollection collection)
+            {
+                if (providers2 != null)
+                {
+                    collection.AddSingleton((Func<IServiceProvider, ILoggerFactory>)delegate (IServiceProvider services)
+                    {
+                        SerilogLoggerFactory serilogLoggerFactory = new SerilogLoggerFactory(logger2, dispose, providers2);
+                        foreach (ILoggerProvider service in services.GetServices<ILoggerProvider>())
+                        {
+                            serilogLoggerFactory.AddProvider(service);
+                        }
+
+                        return serilogLoggerFactory;
+                    });
+                }
+                else
+                {
+                    collection.AddSingleton((Func<IServiceProvider, ILoggerFactory>)((IServiceProvider _) => new SerilogLoggerFactory(logger2, dispose)));
+                }
+
+                ConfigureServices_local(collection, logger2);
+            });
+            return builder;
+        }
+
+        private static void ConfigureServices_local(IServiceCollection collection, Serilog.ILogger? logger)
+        {
+            if (collection == null)
+            {
+                throw new ArgumentNullException("collection");
+            }
+
+            if (logger != null)
+            {
+                collection.AddSingleton(logger);
+            }
+
+            DiagnosticContext implementationInstance = new DiagnosticContext(logger);
+            collection.AddSingleton(implementationInstance);
+            collection.AddSingleton((IDiagnosticContext)implementationInstance);
+        }
+    }
+}
 
 namespace OLT.AspNetCore.Serilog.Tests
 {
@@ -20,8 +89,9 @@ namespace OLT.AspNetCore.Serilog.Tests
         public static IWebHostBuilder WebHostBuilder<T>() where T : class
         {
             var webBuilder = new WebHostBuilder();
+
             webBuilder
-                .UseSerilog()
+                .UseSerilog_local()
                 .ConfigureAppConfiguration(builder =>
                 {
                     builder
@@ -29,10 +99,12 @@ namespace OLT.AspNetCore.Serilog.Tests
                         .AddUserSecrets<T>()
                         .AddJsonFile("appsettings.json", true, false)
                         .AddEnvironmentVariables();
-                })
+                })                
                 .UseStartup<T>();
 
+
             return webBuilder;
+
         }
 
 
