@@ -1,76 +1,76 @@
-using FluentAssertions;
-using Microsoft.Extensions.Options;
-using System;
-using System.IO;
-using System.Linq;
-using System.Text.Json;
-using System.Threading.Tasks;
-using Xunit;
+//using FluentAssertions;
+//using Microsoft.Extensions.Options;
+//using System;
+//using System.IO;
+//using System.Linq;
+//using System.Text.Json;
+//using System.Threading.Tasks;
+//using Xunit;
 
-namespace OLT.Email.SendGrid.Tests
-{
-    public class SendGridConfigurationTests
-    {
-        public class AppSettingsJsonDto
-        {
-            public OltEmailConfigurationSendGrid SendGrid { get; set; } = new OltEmailConfigurationSendGrid();
-        }
+//namespace OLT.Email.SendGrid.Tests
+//{
+//    public class SendGridConfigurationTests
+//    {
+//        public class AppSettingsJsonDto
+//        {
+//            public OltEmailConfigurationSendGrid SendGrid { get; set; } = new OltEmailConfigurationSendGrid();
+//        }
 
 
-        private readonly OltEmailConfigurationSendGrid _emailConfiguration;
+//        private readonly OltEmailConfigurationSendGrid _emailConfiguration;
 
-        public SendGridConfigurationTests(IOptions<OltEmailConfigurationSendGrid> options)
-        {
-            _emailConfiguration = options.Value;
-        }
+//        public SendGridConfigurationTests(IOptions<OltEmailConfigurationSendGrid> options)
+//        {
+//            _emailConfiguration = options.Value;
+//        }
 
-        [Fact]
-        public void EmailConfiguration()
-        {
-            var apiKey = Guid.NewGuid().ToString();
-            var fromEmail = Faker.Internet.Email();
-            var fromName = Faker.Name.FullName();
+//        [Fact]
+//        public void EmailConfiguration()
+//        {
+//            var apiKey = Guid.NewGuid().ToString();
+//            var fromEmail = Faker.Internet.Email();
+//            var fromName = Faker.Name.FullName();
 
-            var whiteEmail = Faker.Internet.Email();
-            var whiteDomain = Faker.Internet.DomainName();
+//            var whiteEmail = Faker.Internet.Email();
+//            var whiteDomain = Faker.Internet.DomainName();
 
-            var config = new OltEmailConfigurationSendGrid();
-            config.ApiKey = apiKey;            
-            config.TestWhitelist.Domain = whiteDomain;
-            config.TestWhitelist.Email = $"{whiteEmail};";
-            config.From.Name = fromName;
-            config.From.Email = fromEmail;
+//            var config = new OltEmailConfigurationSendGrid();
+//            config.ApiKey = apiKey;            
+//            config.TestWhitelist.Domain = whiteDomain;
+//            config.TestWhitelist.Email = $"{whiteEmail};";
+//            config.From.Name = fromName;
+//            config.From.Email = fromEmail;
 
-            Assert.False(config.Production);
-            Assert.Equal(apiKey, config.ApiKey);
-            Assert.Equal(fromName, config.From.Name);
-            Assert.Equal(fromEmail, config.From.Email);
-            Assert.NotEmpty(config.TestWhitelist.Email);
-            Assert.NotEmpty(config.TestWhitelist.Domain);
-            Assert.Equal(whiteEmail, config.TestWhitelist.EmailParsed.ToList()[0]);
-            Assert.Equal(whiteDomain, config.TestWhitelist.DomainParsed.ToList()[0]);
+//            Assert.False(config.Production);
+//            Assert.Equal(apiKey, config.ApiKey);
+//            Assert.Equal(fromName, config.From.Name);
+//            Assert.Equal(fromEmail, config.From.Email);
+//            Assert.NotEmpty(config.TestWhitelist.Email);
+//            Assert.NotEmpty(config.TestWhitelist.Domain);
+//            Assert.Equal(whiteEmail, config.TestWhitelist.EmailParsed.ToList()[0]);
+//            Assert.Equal(whiteDomain, config.TestWhitelist.DomainParsed.ToList()[0]);
 
-            config.Production = true;
-            Assert.True(config.Production);
+//            config.Production = true;
+//            Assert.True(config.Production);
 
-        }
+//        }
 
-        [Fact]
-        public async Task Options()
-        {
-            Assert.NotNull(_emailConfiguration);
+//        [Fact]
+//        public async Task Options()
+//        {
+//            Assert.NotNull(_emailConfiguration);
 
-            string fileName = "appsettings.json";
-            var filePath = Path.Combine(AppContext.BaseDirectory, fileName);
+//            string fileName = "appsettings.json";
+//            var filePath = Path.Combine(AppContext.BaseDirectory, fileName);
 
-            using (FileStream openStream = File.OpenRead(filePath))
-            {
-                AppSettingsJsonDto expectedConfig = await JsonSerializer.DeserializeAsync<AppSettingsJsonDto>(openStream);
-                Assert.NotNull(expectedConfig);
-                _emailConfiguration.Should().BeEquivalentTo(expectedConfig?.SendGrid);
-            }
+//            using (FileStream openStream = File.OpenRead(filePath))
+//            {
+//                AppSettingsJsonDto expectedConfig = await JsonSerializer.DeserializeAsync<AppSettingsJsonDto>(openStream);
+//                Assert.NotNull(expectedConfig);
+//                _emailConfiguration.Should().BeEquivalentTo(expectedConfig?.SendGrid);
+//            }
 
-        }
+//        }
 
-    }
-}
+//    }
+//}
