@@ -3,9 +3,11 @@ using System.Collections.Generic;
 using Xunit;
 using OLT.Core;
 using System;
+using Newtonsoft.Json.Linq;
 
 namespace OLT.Extensions.Configuration.Tests
 {
+
     public class ConnectionStringTests
     {
         readonly string name = "DBConnection";
@@ -29,16 +31,13 @@ namespace OLT.Extensions.Configuration.Tests
 
         [Fact]
         public void OverrideTest()
-        {
-            
+        {            
 
             var myConfiguration = new Dictionary<string, string>
             {
                 {$"connection-strings:{name}", "Value1"},
                 {$"ConnectionStrings:{name}", "Value2"}
-            };
-
-           
+            };           
 
             var configuration = new ConfigurationBuilder()
                 .AddEnvironmentVariables()
@@ -52,7 +51,6 @@ namespace OLT.Extensions.Configuration.Tests
         [Fact]
         public void StandardTest()
         {
-
             var myConfiguration = new Dictionary<string, string>
             {
                 {$"ConnectionStrings:{name}", "Value2"}
